+++ ProiectDAW/Controllers/PostsController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Identity;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.VisualStudio.Web.CodeGenerators.Mvc.Templates.BlazorIdentity.Pages;
@@ -31,9 +32,10 @@ namespace ProiectDAW.Controllers
 
             ViewBag.Posts = posts;
             //afisez mesajul din delete
-            if (TempData.ContainsKey("message"))
+            if (TempData.ContainsKey("msg"))
             {
-                ViewBag.Msg = TempData["message"];
+                ViewBag.Msg = TempData["msg"];
+                ViewBag.MsgType = TempData["msgType"];
             }
             return View();
         }
@@ -41,14 +43,35 @@ namespace ProiectDAW.Controllers
         public IActionResult Show(int Id)
         {
             Post post = db.Posts.Include("Comments")
+                                .Include("User")
+                                .Include("Comments.User")    
                                 .Where(post => post.Id == Id)
                                 .First();
+            ViewBag.Buttons = false;
+            if (User.Identity.IsAuthenticated)
+            {
+                ViewBag.Buttons = true;
+            }
+
             return View(post);
         }
         [HttpPost]
+        [Authorize(Roles = "User,Admin")]
         public IActionResult Show([FromForm] Comment comment) //adugare comentarii
         {
+            if (!User.Identity.IsAuthenticated)
+            {
+                string returnUrl = HttpContext.Request.Path; //calea catre pagina curenta
+                string loginUrl = $"/Identity/Account/Login?returnUrl={Uri.EscapeDataString(returnUrl)}"; //calea catre pagina de login
+                Console.WriteLine(returnUrl.ToString());
+                return Redirect(loginUrl);
+            }
+
+
             comment.Date = DateTime.Now;
+            
+            // preluam Id-ul utilizatorului care posteaza comentariul
+            comment.UserId = _userManager.GetUserId(User);
             if (ModelState.IsValid)
             {
                 db.Comments.Add(comment);
@@ -66,18 +89,24 @@ namespace ProiectDAW.Controllers
             }
         }
 
+        [Authorize(Roles = "User,Admin")]
         public IActionResult New()
         {
             if (!User.Identity.IsAuthenticated)
             {
-                return Redirect("/Identity/Account/Login"); // Controllerul și metoda Login
+                string returnUrl = HttpContext.Request.Path; //calea catre pagina curenta
+                string loginUrl = $"/Identity/Account/Login?returnUrl={Uri.EscapeDataString(returnUrl)}"; //calea catre pagina de login
+                Console.WriteLine(returnUrl.ToString());
+                return Redirect(loginUrl);
             }
 
             return View();
         }
+        [Authorize(Roles = "User,Admin")]
         [HttpPost]
         public async Task<IActionResult> New(Post post, IFormFile Image)
         {
+
             if (Image != null && Image.Length > 0)
             {
                 // Verificăm extensia
@@ -99,33 +128,54 @@ namespace ProiectDAW.Controllers
                 {
                     await Image.CopyToAsync(fileStream);
                 }
+                
                 ModelState.Remove(nameof(post.Image));
                 post.Image = databaseFileName;
             }
 
-            try
+            ModelState.Remove(nameof(post.Image));   //elimina eroarea de validare a pozei
+
+            if (string.IsNullOrEmpty(post.Description) && (Image == null || Image.Length == 0))
             {
-                post.Date = DateTime.Now;
-                post.Likes = 0;
-                post.UserId = _userManager.GetUserId(User);
+                ModelState.AddModelError("", "Cel puțin un câmp trebuie să fie completat (Description sau Image).");
+                return View(post);
+            }
+
+            post.Date = DateTime.Now;
+            post.Likes = 0;
+            post.UserId = _userManager.GetUserId(User);
+            
+            if(ModelState.IsValid)
+            {
+                
                 db.Posts.Add(post);
                 db.SaveChanges();
+
+                TempData["msg"] = "Postarea a fost adaugata!";
+                TempData["msgType"] = "alert-success"; //clasa de bootstrap pt mesaj de succes(verde)
                 return RedirectToAction("Index");
             }
-            catch (Exception ex)
+            else
             {
-                Console.WriteLine($"Eroare la adăugarea postării: {ex.Message}");
-                Console.WriteLine($"StackTrace: {ex.StackTrace}");
                 // Dacă nu este autentificat, redirecționează la Login
-                
                 return View(post);
             }
         }
-
-        public IActionResult Edit(int Id)
+        [Authorize(Roles = "User,Admin")]
+        public IActionResult Edit(int Id) //doar userul care a postat poate edita
         {
             Post post = db.Posts.Find(Id);
-            return View(post);
+
+            if (post.UserId == _userManager.GetUserId(User)) //verificam daca userul curent este cel care a postat
+            {                                                //User.IsInRole("Admin") - verifica daca userul este admin
+                return View(post);
+            }
+            else
+            {
+                TempData["msg"] = "Nu aveți dreptul să editați această postare!";
+                TempData["msgType"] = "alert-danger"; //clasa de bootstrap pt mesaj de eroare(rosu)
+                return RedirectToAction("Index");
+            }
         }
         [HttpPost]
         public async Task<IActionResult> Edit(int Id, Post editedPost, IFormFile Image)
@@ -156,31 +206,77 @@ namespace ProiectDAW.Controllers
                 ModelState.Remove(nameof(post.Image));
                 post.Image = databaseFileName;
             }
+    
 
-            try
+            ModelState.Remove(nameof(post.Image));
+            //verififcam daca cel putin un camp este completat (sau daca posarea are deja imagine sau descriere)
+            if (string.IsNullOrEmpty(editedPost.Description) && string.IsNullOrEmpty(post.Description) && (Image == null || Image.Length == 0) && (post.Image == null || post.Image.Length == 0))
             {
-                post.Description = editedPost.Description;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                ModelState.AddModelError("", "Cel puțin un câmp trebuie să fie completat (Description sau Image).");
+                return View(editedPost);
             }
-            catch (Exception)
+
+            if (ModelState.IsValid)
             {
-                return RedirectToAction("Edit", post.Id);
+                if (post.UserId == _userManager.GetUserId(User)) //verificam daca userul curent este cel care a postat
+                {                                                //User.IsInRole("Admin") - verifica daca userul este admin
+                    post.Description = editedPost.Description;
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+                else
+                {
+                    TempData["msg"] = "Nu aveți dreptul să editați această postare!";
+                    TempData["msgType"] = "alert-danger"; //clasa de bootstrap pt mesaj de eroare(rosu)
+                    return RedirectToAction("Index");
+                }
+
+            }
+            else
+            {
+                return View(editedPost); //editedPost ca a se pastreze datele introduse
             }
         }
+        
+        public IActionResult DeleteImage(int Id)
+        {
 
+            Post post = db.Posts.Find(Id);
+            if (post.Image != null)
+            {
+                post.Image = null;
+                db.SaveChanges();
+                Console.WriteLine("Imaginea a fost stearsa!");
+            }
+            else
+            {
+                Console.WriteLine("Nu exista imagine de sters!");
+            }
+            return View("Edit", post);
+        }
+
+        [Authorize(Roles = "User,Admin")]
         [HttpPost]
         public IActionResult Delete(int Id)
         {
             Post post = db.Posts.Include(p => p.Comments).FirstOrDefault(p => p.Id == Id);
-            
-            //stergem comentariile postarii
-            db.Comments.RemoveRange(post.Comments);
 
-            db.Posts.Remove(post);
-            db.SaveChanges();
+            if (post.UserId == _userManager.GetUserId(User))
+            {
+                //stergem comentariile postarii
+                db.Comments.RemoveRange(post.Comments);
 
-            TempData["message"] = "Postarea a fost stearsa!";
+                db.Posts.Remove(post);
+                db.SaveChanges();
+
+                TempData["msg"] = "Postarea a fost stearsa!";
+                TempData["msgType"] = "alert-danger"; //clasa de bootstrap pt mesaj de eroare(rosu)
+            }
+            else
+            {
+                TempData["msg"] = "Nu aveți dreptul să ștergeți această postare!";
+                TempData["msgType"] = "alert-danger"; //clasa de bootstrap pt mesaj de eroare(rosu)
+            }
             return RedirectToAction("Index");
         }
     }
